requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
jinja2==3.1.4
transformers==4.40.2
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
requests==2.31.0
numpy==1.26.4
python-dotenv==1.0.1
